refactor: package to not have barrels
