--- packages/cache/src/cache/index.ts
@@ -1,9 +1,0 @@
-const version = '__@FLOPFLIP/VERSION_OF_RELEASE__';
-
-export { version };
-export {
-  encodeCacheContext,
-  getAllCachedFlags,
-  getCache,
-  getCachedFlags,
-} from './cache';
